+++ include/boost/multiprecision/detail/generic_interconvert.hpp
@@ -136,17 +136,17 @@ void generic_interconvert(To& to, const From& from, const mpl::int_<number_kind_
    //
    int c = eval_fpclassify(from);
 
-   if(c == FP_ZERO) 
+   if(c == (int)FP_ZERO)
    {
       to = ui_type(0);
       return;
    }
-   else if(c == FP_NAN)
+   else if(c == (int)FP_NAN)
    {
       to = static_cast<const char*>("nan");
       return;
    }
-   else if(c == FP_INFINITE)
+   else if(c == (int)FP_INFINITE)
    {
       to = static_cast<const char*>("inf");
       if(eval_get_sign(from) < 0)
